+++ src/main/java/core/nmvc/scan/ClasspathBeanDefinitionScanner.java
@@ -1,4 +1,4 @@
-package core.nmvc;
+package core.nmvc.scan;
 
 import com.google.common.collect.Sets;
 import core.annotation.Controller;
@@ -9,11 +9,11 @@ import org.reflections.Reflections;
 import java.lang.annotation.Annotation;
 import java.util.Set;
 
-public class BeanScanner {
+public class ClasspathBeanDefinitionScanner {
 
   private final Reflections reflections;
 
-  public BeanScanner(Object... basePackage) {
+  public ClasspathBeanDefinitionScanner(Object... basePackage) {
     // ! 중요한거,  Reflections 의 생성자로 파라미터가 없는 경우 ([] length == 0 인 경우) 조건이 없는 경우 모든 패키지에서 찾는다.
     this.reflections = new Reflections(basePackage);
   }
